requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
black==21.5b2
bleach==3.3.0
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.5
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
emoji==1.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
entrypoints==0.3
flake8==3.9.2
idna==2.10
iniconfig==1.1.1
install==1.3.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
mail-parser==3.15.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
mypy-extensions==0.4.3
nameparser==1.0.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
preshed==3.0.5
progressbar2==3.53.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-cov==2.12.1
python-dateutil==2.8.1
python-slugify==5.0.2
python-utils==2.5.6
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.16.0
smart-open==5.1.0
soupsieve==2.2.1
spacy==3.1.0
spacy-legacy==3.0.8
srsly==2.4.1
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
thinc==8.0.7
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xgboost==1.4.2
